# docs/swagger.json
{"openapi":"3.0.1","info":{"title":"Chest X-ray Diagnosis API","description":"AI 기반 폐 질환 진단 서비스 API 문서","version":"v1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Auth","description":"Firebase 로그인 API"},{"name":"Chest X-ray Diagnosis","description":"AI 진단 기능 관련 API"}],"paths":{"/api/v1/diagnosis/upload":{"post":{"tags":["Chest X-ray Diagnosis"],"summary":"X-ray 업로드","description":"DICOM 또는 이미지 파일을 업로드하여 AI 분석 요청","operationId":"uploadXray","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","description":"X-ray 이미지 파일 (jpg/png/dcm 등)","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Auth"],"summary":"이메일/비밀번호 회원가입","description":"Firebase로 회원가입을 수행하고 ID 토큰을 반환합니다.","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"이메일/비밀번호 로그인","description":"Firebase Auth로 로그인 후 ID 토큰을 반환합니다.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/news/add":{"post":{"tags":["news-controller"],"operationId":"addNews","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/News"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/News"}}}}}}},"/api/v1/diagnosis/result":{"get":{"tags":["Chest X-ray Diagnosis"],"summary":"AI 진단 결과","description":"업로드된 X-ray에 대한 AI 분석 결과를 반환","operationId":"getResult","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/news":{"get":{"tags":["news-controller"],"operationId":"getAllNews","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/News"}}}}}}}}},"components":{"schemas":{"SignupRequestDto":{"type":"object","properties":{"email":{"type":"string","description":"사용자 이메일","example":"test@example.com"},"password":{"type":"string","description":"비밀번호","example":"yourPassword123"},"nickname":{"type":"string","description":"닉네임","example":"홍길동"}}},"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string","description":"사용자 이메일 (로그인 ID)","example":"test@example.com"},"password":{"type":"string","description":"비밀번호","example":"yourPassword123"}}},"News":{"type":"object","properties":{"newsId":{"type":"integer","format":"int32"},"title":{"type":"string"},"newsUrl":{"type":"string"},"summary":{"type":"string"},"publishedAt":{"type":"string","format":"date-time"},"crawledAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}